// src/app/interfaces/about.ts
// export interface About {
//   btnLink?: string;
//   btnText?: string;
//   desc?: string;
//   imageUrl?: string;
//   subtitle?: string;
//   title?: string;
//   titleTwo?: string;
// }

// export interface AboutCta {
//   btnLink?: string;
//   btnText?: string;
//   desc?: string;
//   title?: string;
// }

// export interface Qualification {
//   title?: string;
// }

export interface Gallery {
  imageUrl?: string;
}

export interface About {
  id?: string;
  title?: string;
  titleSpan?: string;
  subtitle?: string;
  desc?: string;
  para1?: string;
  para2?: string;
  para3?: string;
  btnText?: string;
  btnLink?: string;
  imageUrl?: string;
  titleTwo?: string;
  titleTwoSpan?: string;
  qualifications?: any;
}

export interface AboutCta {
  id?: string;
  title?: string;
  titleSpan?: string;
  desc?: string;
  para?: string;
  btnText?: string;
  btnLink?: string;
}

export interface Qualification {
  id?: string;
  title?: string;
}